+++ src/ng2-ue-utils.module.ts
@@ -1,22 +1,26 @@
-import { CommonModule }         from '@angular/common';
-import { NgModule }             from '@angular/core';
-import { Ng2BootstrapModule }   from 'ng2-bootstrap/ng2-bootstrap';
+import { CommonModule }             from '@angular/common';
+import { FormsModule }              from '@angular/forms';
+import { NgModule }                 from '@angular/core';
 
-import { Filtering }            from './components/filtering/filtering';
-import { HintScroll }           from './components/hint-scroll/hint-scroll';
-import { ListHeaders }          from './components/list-headers/list-headers';
-import { LoadingDots }          from './components/loading-dots/loading-dots';
-import { Modal }                from './components/modal/modal';
-import { Notification }         from './components/notification/notification';
-import { Pagination }           from './components/pagination/pagination';
-import { Progress }             from './components/progress/progress';
-import { ValueHandler }         from './components/value-handler/value-handler';
-import { FocusOnInit }          from './directives/focus-on-init/focus-on-init';
-import { ToggleOnHover }        from './directives/toggle-on-hover/toggle-on-hover';
-import { DateMoment }           from './pipes/date-moment/date-moment';
-import { FromNow }              from './pipes/from-now/from-now';
-import { DatePicker }           from './components/datepicker/datepicker';
-import { DatePickerInput }      from './components/datepicker-input/datepicker-input';
+import { Filtering }                from './components/filtering/filtering';
+import { HintScroll }               from './components/hint-scroll/hint-scroll';
+import { ListHeaders }              from './components/list-headers/list-headers';
+import { LoadingDots }              from './components/loading-dots/loading-dots';
+import { Modal }                    from './components/modal/modal';
+import { Notification }             from './components/notification/notification';
+import { Pagination }               from './components/pagination/pagination';
+import { Progress }                 from './components/progress/progress';
+import { ValueHandler }             from './components/value-handler/value-handler';
+import { Tooltip }                  from './components/tooltip/tooltip';
+import { FocusOnInit }              from './directives/focus-on-init/focus-on-init';
+import { ToggleOnHover }            from './directives/toggle-on-hover/toggle-on-hover';
+import { TooltipDirective }         from './directives/tooltip/tooltip';
+import { DateMoment }               from './pipes/date-moment/date-moment';
+import { FromNow }                  from './pipes/from-now/from-now';
+import { DatePicker }               from './components/datepicker/datepicker';
+import { DatePickerInput }          from './components/datepicker-input/datepicker-input';
+import { DropdownDirective }        from './directives/dropdown/dropdown';
+import { DropdownTriggerDirective } from './directives/dropdown/dropdown-trigger';
 
 @NgModule({
     declarations: [
@@ -31,6 +35,10 @@ import { DatePickerInput }      from './components/datepicker-input/datepicker-i
         ValueHandler,
         FocusOnInit,
         ToggleOnHover,
+        Tooltip,
+        TooltipDirective,
+        DropdownDirective,
+        DropdownTriggerDirective,
         DateMoment,
         FromNow,
         DatePicker,
@@ -48,6 +56,10 @@ import { DatePickerInput }      from './components/datepicker-input/datepicker-i
         ValueHandler,
         FocusOnInit,
         ToggleOnHover,
+        Tooltip,
+        TooltipDirective,
+        DropdownTriggerDirective,
+        DropdownDirective,
         DateMoment,
         FromNow,
         DatePicker,
@@ -55,7 +67,8 @@ import { DatePickerInput }      from './components/datepicker-input/datepicker-i
     ],
     imports: [
         CommonModule,
-        Ng2BootstrapModule
+        FormsModule
     ]
+
 })
 export class Ng2UEUtilsModule {}